backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
pandas>=2.2.0
chardet>=5.2.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.7.0
numpy>=1.26.0
